Fixed items dropping off players with bad connection Reduced vector3 and quaternion network size by using shorts instead of floats

+++ Multiplayer Mod/Multiplayer Mod/Packets/Packet.cs
@@ -124,17 +124,17 @@ namespace Multiplayer_Mod
 
 		public void Write(Vector3 _value)
 		{
-			Write(_value.x);
-			Write(_value.y);
-			Write(_value.z);
+			Write((short)_value.x);
+			Write((short)_value.y);
+			Write((short)_value.z);
 		}
 
 		public void Write(Quaternion _value)
 		{
-			Write(_value.x);
-			Write(_value.y);
-			Write(_value.z);
-			Write(_value.w);
+			Write((short)_value.x);
+			Write((short)_value.y);
+			Write((short)_value.z);
+			Write((short)_value.w);
 		}
 
 		public void Write(Color _value)
